# swagger.yaml
swagger: "2.0"
info:
  description: " Some description  "
  version: "1.0.0"
  title: "Some title for API Documentation"
host: "example-app"
basePath: "/public/api/v1"
...
schemes:
- "http"
paths:
  /users:
    post:
      tags:
      - "users"
      summary: "Add a new user"
      description: ""
      operationId: "addUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "status ok"
        "400":
          description: "such a user exists"
    get:
      tags:
      - "users"
      summary: "Get all users"
      description: ""
      operationId: "getUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "status ok"
...
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
...